requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
mysql-connector-python==8.3.0
PyJWT==2.8.0
bcrypt==4.1.2
python-multipart==0.0.6
pydantic[email]==2.6.0
python-dotenv==1.0.1
aiosmtplib==3.0.1
slowapi==0.1.9
limits==3.7.0
cryptography==42.0.5
gunicorn==21.2.0